[+] update: wrap try-catch

+++ src/routers/deleteTransectionRouter.ts
@@ -5,12 +5,19 @@ const router = express.Router();
 
 router.delete("/api/transection/:id", async (req: Request, res: Response) => {
   const { id } = req.params;
+  if (id === "") {
+    return res.json({ message: "id is required" });
+  }
 
-  const data = await client.query(
-    "delete from transection where transection_id=$1 returning *",
-    [id]
-  );
-  res.json({ message: "delete success", data: data.rows });
+  try {
+    const data = await client.query(
+      "delete from transection where transection_id=$1 returning *",
+      [id]
+    );
+    return res.json({ message: "delete success", data: data.rows });
+  } catch (error) {
+    return res.status(500).json({ message: "Internal Server Error" });
+  }
 });
 
 export { router as deleteTransectionRouter };
